Support default and floating point rates for traffic shaping.

+++ svc/dappvpn/data/offering.go
@@ -2,6 +2,6 @@ package data
 
 // OfferingParams is an offering additional parameters for VPN.
 type OfferingParams struct {
-	MinUploadMbits   int `json:"minUploadMbits"`
-	MinDownloadMbits int `json:"minDownloadMbits"`
+	MinUploadMbits   float32 `json:"minUploadMbits"`
+	MinDownloadMbits float32 `json:"minDownloadMbits"`
 }
