+++ tc/tc_windows.go
@@ -12,7 +12,7 @@ func NewConfig() *Config {
 // SetRateLimit sets a rate limit for a given client IP address on a given
 // network interface.
 func (tc *TrafficControl) SetRateLimit(
-	iface, clientIP string, upMbps, downMbps int) error {
+	iface, clientIP string, upMbps, downMbps float32) error {
 	return nil
 }
